requirements.txt
pyrogram==1.4.16
TgCrypto
py-tgcalls==0.8.6
python-dotenv
youtube_search_python
requests
yt-dlp==2021.12.01
aiohttp
aiofiles
asyncio
python-arq
youtube_search
search_engine_parser
ffmpeg-python
ffmpeg
Pillow
ujson
psutil
telegram
python-telegram-bot==13.11
